web_app/backend/requirements.txt
fastapi
uvicorn
python-multipart
rasterio
numpy
torch
torchvision
pillow
matplotlib
